Changed movement logic
In order to solve strange motion behaviour on high velocities, the
motion logic was put into separate functions.
The horizontal movement updates the vertical velocity on first contact
as before, just the position of the lines in the code changed.

Will do another branch, because I wanted a do/while loop and it would be
very similar to the attached movement...